pyproject.toml
[project]
name = "briehost-api"
version = "0.1.0"
description = "Upload + Proxmox LXC provisioning API for BrieHosting"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "pyjwt[crypto]>=2.9",
    "httpx>=0.27",
    "supabase>=2.7",
    "proxmoxer>=2.0",
    "requests>=2.32",
    "python-dotenv>=1.0",
]

[tool.uv]
dev-dependencies = ["ruff>=0.6"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
